requirements.txt
Django==2.1.2
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.9.0
django-debug-toolbar==1.8
whitenoise==3.3.1
requests==2.20.0
simplejson==3.16.0
django-cors-headers==2.4.0
gunicorn==19.9.0
psycopg2-binary==2.7.5
